requirements.txt
django==4.2
pillow
django-simple-captcha